add resolvePuppet to export
