+++ src/mods/helpers.ts
@@ -1,0 +1,1 @@
+export { resolvePuppet } from '../puppet-management/mod.js'
