+++ src/puppet-management/mod.ts
@@ -1,23 +1,17 @@
-import type {
-  PuppetModuleName,
-}                         from './puppet-config.js'
 import {
+  type PuppetModuleName,
   PUPPET_DEPENDENCIES,
   PUPPET_NAME_DEFAULT,
 }                         from './puppet-config.js'
-import type {
-  ResolveOptions,
-}                         from './puppet-manager.js'
 import {
-  PuppetManager,
-}                         from './puppet-manager.js'
+  type ResolveOptions,
+  resolvePuppet,
+}                         from './puppet-resolver.js'
 
-export type {
-  PuppetModuleName,
-  ResolveOptions,
-}
 export {
+  type PuppetModuleName,
+  type ResolveOptions,
   PUPPET_DEPENDENCIES,
   PUPPET_NAME_DEFAULT,
-  PuppetManager,
+  resolvePuppet,
 }
